refactor(agglayer-types): route SP1 access through agglayer-sp1 facade
Remove the direct sp1-sdk, sp1-sdk-v5, sp1-core-machine and sp1-prover
dependencies from agglayer-types. The test-only mock-proof builder
dummy_sp1_stark_proof_with_version moves into agglayer-sp1::testutils
and is re-exported, so all consumers are unchanged. sp1-core-machine
and sp1-prover were already unused.

Refs: #1524
