+++ crates/agglayer-types/src/certificate/testutils.rs
@@ -1,6 +1,5 @@
 use agglayer_interop_types::{aggchain_proof::AggchainData, LocalExitRoot};
 use agglayer_primitives::{Address, Digest, Hashable, Signature, B256};
-use agglayer_sp1::testutils::{EMPTY_ELF, EMPTY_ELF_V5};
 use pessimistic_proof::{
     core::commitment::{SignatureCommitmentValues, SignatureCommitmentVersion},
     keccak::keccak256_combine,
@@ -245,85 +244,10 @@ pub enum AggchainDataType {
     MultisigAndAggchainProof { num_signers: usize },
 }
 
-fn sp1_stark_with_context<P: serde::Serialize, V: serde::Serialize>(
-    proof: &P,
-    vkey: &V,
-    version: &str,
-) -> agglayer_interop_types::aggchain_proof::SP1StarkWithContext {
-    agglayer_interop_types::aggchain_proof::SP1StarkWithContext {
-        proof: agglayer_interop_types::bincode::default()
-            .serialize(proof)
-            .unwrap(),
-        vkey: agglayer_interop_types::bincode::default()
-            .serialize(vkey)
-            .unwrap(),
-        version: version.to_owned(),
-    }
-}
-
-/// Create a dummy STARK proof for testing purposes with a specific SP1 version.
-pub fn dummy_sp1_stark_proof_with_version(
-    version: &str,
-) -> agglayer_interop_types::aggchain_proof::Proof {
-    match agglayer_sp1::version_kind(version) {
-        Ok(agglayer_sp1::Sp1ProofVersion::V5) => {
-            use sp1_sdk_v5::Prover;
-
-            // The legacy v5 mock prover spins up its own tokio runtime during
-            // construction, so build it on a fresh OS thread to avoid nesting
-            // runtimes inside async tests.
-            std::thread::spawn({
-                let version = version.to_owned();
-                move || {
-                    let client = sp1_sdk_v5::ProverClient::builder().mock().build();
-                    let (proving_key, vkey) = client.setup(EMPTY_ELF_V5);
-                    let dummy_proof = sp1_sdk_v5::SP1ProofWithPublicValues::create_mock_proof(
-                        &proving_key,
-                        sp1_sdk_v5::SP1PublicValues::new(),
-                        sp1_sdk_v5::SP1ProofMode::Compressed,
-                        sp1_sdk_v5::SP1_CIRCUIT_VERSION,
-                    );
-                    let proof = dummy_proof.proof.try_as_compressed().unwrap();
-
-                    agglayer_interop_types::aggchain_proof::Proof::SP1Stark(sp1_stark_with_context(
-                        proof.as_ref(),
-                        &vkey,
-                        &version,
-                    ))
-                }
-            })
-            .join()
-            .expect("legacy v5 mock-proof thread should not panic")
-        }
-        _ => {
-            use agglayer_sp1::{v6_sp1_stark_with_context, V6Sp1StarkProof};
-            use sp1_sdk::{blocking::Prover, ProvingKey};
-
-            let (proof, vkey) = {
-                let client = sp1_sdk::blocking::ProverClient::builder().mock().build();
-                let proving_key = client.setup(sp1_sdk::Elf::Static(EMPTY_ELF)).unwrap();
-                let verif_key = proving_key.verifying_key().clone();
-                let dummy_proof = sp1_sdk::SP1ProofWithPublicValues::create_mock_proof(
-                    proving_key.verifying_key(),
-                    sp1_sdk::SP1PublicValues::new(),
-                    sp1_sdk::SP1ProofMode::Compressed,
-                    sp1_sdk::SP1_CIRCUIT_VERSION,
-                );
-                let proof: Box<V6Sp1StarkProof> = dummy_proof.proof.try_as_compressed().unwrap();
-                (proof, verif_key)
-            };
-
-            let mut sp1 = v6_sp1_stark_with_context(proof.as_ref(), &vkey, "v6.1.0").unwrap();
-            sp1.version = version.to_owned();
-            agglayer_interop_types::aggchain_proof::Proof::SP1Stark(sp1)
-        }
-    }
-}
-
 /// Create a dummy STARK proof for testing purposes.
 /// This creates a minimal SP1 proof that can be used in tests.
 fn create_dummy_stark_proof() -> agglayer_interop_types::aggchain_proof::Proof {
-    dummy_sp1_stark_proof_with_version("test")
+    agglayer_sp1::testutils::dummy_sp1_stark_proof_with_version("test")
 }
 
 impl Certificate {
